requirements.txt
Django==1.6.2
Pillow==2.4.0
South==0.8.4
argparse==1.2.1
boto==2.27.0
dj-database-url==0.3.0
dj-static==0.0.5
django-dajax==0.9.2
django-dajaxice==0.6
django-debug-toolbar==1.1
django-embed-video==0.9
django-extensions==1.3.3
django-storages==1.1.8
django-tastypie==0.11.1
django-toolbelt==0.0.1
gunicorn==18.0
ipython==2.0.0
psycopg2==2.5.2
pystache==0.5.3
python-dateutil==2.2
python-mimeparse==0.1.4
requests==2.2.1
six==1.6.1
sorl-thumbnail==11.12
sqlparse==0.1.11
static==1.0.2
wsgiref==0.1.2
